Enable building with GCC 14
When I tried to build the package with GCC 14, I got the following
error:

    src/ThirdParty/Minizip/minizip/miniunz.c:141:11: error:
    implicit declaration of function ‘mkdir’; did you mean ‘mymkdir’?
    [-Wimplicit-function-declaration]

It turns out that GCC now conforms more strictly to the C standard, and
some issues that were previously only warned about are now treated as
errors.  `implicit-function-declaration` is one of these.  For more
info, see:
https://gcc.gnu.org/gcc-14/porting_to.html#implicit-function-declaration

The solution is to patch the Minizip code so that the appropriate header
for `mkdir()`, which is `<sys/stat.h>`, gets included.